requirements.txt
alembic==1.8.1
astroid==2.2.5
Click==7.0
colorama==0.4.5
ecdsa==0.13.2
Flask==1.0.2
Flask-Cors==3.0.8
Flask-Migrate==2.7.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.5.0
future==0.17.1
greenlet==1.1.3
gunicorn==20.1.0
importlib-metadata==4.12.0
importlib-resources==5.10.0
isort==4.3.18
itsdangerous==1.1.0
Jinja2==2.10.1
lazy-object-proxy==1.4.0
Mako==1.2.3
MarkupSafe==1.1.1
mccabe==0.6.1
psycopg2==2.9.5
psycopg2-binary==2.9.1
pyasn1==0.4.8
pycryptodome==3.9.9
pylint==2.3.1
python-dotenv==0.21.0
python-jose==3.3.0
rsa==4.9
six==1.12.0
SQLAlchemy==1.4.41
typed-ast==1.4.2
typing-extensions==4.3.0
Werkzeug==0.16.1
wrapt==1.11.1
zipp==3.8.1
